+++ NEWS.md
@@ -6,6 +6,10 @@
 * Added support for using PhyloPic silhouettes as vertices when plotting `{igraph}` networks via a new `"phylopic"` vertex shape, registered automatically when both packages are loaded (#115, #118)
 * Added new "Network plots" sections to both advanced vignettes, demonstrating the new `{igraph}` integration in base R and the use of `geom_phylopic()` inside `{ggraph}` plots
 
+Deprecation:
+
+* The "ysize" and "size" arguments/aesthetics are now fully deprecated in favor of "height" and "width" arguments/aesthetics. These arguments/aesthetics will be removed in a future version of rphylopic.
+
 # rphylopic 1.6.0
 
 * Added `add_phylopic_tree()` to add silhouettes to base R trees (#110)